pyproject.toml
[tool.poetry]
name = "artist-feedback-bot"
version = "0.1.0"
description = ""
authors = ["whoisridze <boasting_mashing830@simplelogin.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13.2"
pyTelegramBotAPI = "4.26.0"
python-dotenv = "1.0.1"
loguru = "^0.7.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"